:bug: always emit a response
When the local server does not return a valid JSON body, the request times out since no response is sent back to the Jovo Webhook. The logged stack trace is not very helpful as it always just says: "Something went wrong".

With this change the invalid response body is logged instead and an event with a null response is emitted via the socket connection to trigger an immediate output.